+++ source/Obstacles/ZebraFish.cpp
@@ -328,7 +328,7 @@ ZebraFish::ZebraFish(SimulationData&s, ArgumentParser&p, Real C[2]): Fish(s,p,C)
 {
     const Real ampFac = p("-amplitudeFactor").asDouble(1.0);
     myFish = new BehaviorCurvatureFish(length, Tperiod, phaseShift, sim.minH, ampFac);
-    if( sim.rank == 0 && s.verbose ) printf("[CUP2D] - BehaviorCurvatureFish %d %f %f %f\n",myFish->Nm, length, Tperiod, phaseShift);
+    if( sim.rank == 0 && s.verbose ) printf("[CUP2D] - BehaviorCurvatureFish %d %f %f %f\n",myFish->Nm, (double)length, (double)Tperiod, (double) phaseShift);
 }
 
 void ZebraFish::resetAll() {
